+++ api/core/v2/check_config.go
@@ -116,6 +116,12 @@ func (c *CheckConfig) Validate() error {
 		}
 	}
 
+	for _, subscription := range c.Subscriptions {
+		if subscription == "" {
+			return fmt.Errorf("subscriptions cannot be empty strings")
+		}
+	}
+
 	// The entity can be empty but can't contain invalid characters (only
 	// alphanumeric string)
 	if c.ProxyEntityName != "" {